Problem with ending the game
I ran into a problem with having the dealer draw and then ending the game. This is mainly due to the inability to directly chose where to insert html code inside of php. I plan to change the buttons I created in php and use javascript instead and use PHP for all backend work. This will give me the opportunity to grow my skills with javascript.